Integrate modernize_target improvements and fix all warnings
- Sidestep-Info.plist: CFBundleIdentifier changed from
  com.faisal.${PRODUCT_NAME:rfc1034identifier} to $(PRODUCT_BUNDLE_IDENTIFIER)
  (canonical modern Xcode form; picks up value from project build settings)
- AppUtilities.m: removed ~30 lines of dead commented-out timing/elapsed-time
  infrastructure left over from 2010; _XLog is now clean production code
- AppUtilities.h/.m: bool → BOOL on existsInArray: (Cocoa type convention);
  CFIndex for buflen to avoid 64→32 truncation; explicit (size_t) cast for malloc
- DefaultsController.h/.m: remove all [defaults synchronize] calls (unnecessary
  since macOS 10.12; framework flushes automatically); saveSSHConnectionPID:/
  getSSHConnectionPID return type int → NSInteger to match NSUserDefaults API
- EMKeychainItem.m: filter errSecItemNotFound from error log; looking up a
  missing keychain item is normal (first launch, no stored password), not an error
- SSHConnector.h/.m: killSSHConnectionForPID: int → NSInteger to match callers;
  explicit (pid_t) cast at kill() call site
- AppController.m: previousPID int → NSInteger to match updated DefaultsController

+++ AppController.m
@@ -145,7 +145,7 @@ NSInteger GrowlSpam_TestConnection    = 0;
 
     [growl requestAuthorization];
 
-    int previousPID = [defaultsController getSSHConnectionPID];
+    NSInteger previousPID = [defaultsController getSSHConnectionPID];
 
     if (previousPID != 0) {
         XLog(self, @"Turning proxy off");
